+++ src/types/geometry_type.rs
@@ -1,0 +1,11 @@
+/// The geometry type of the WKT object
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+pub enum GeometryType {
+    Point,
+    LineString,
+    Polygon,
+    MultiPoint,
+    MultiLineString,
+    MultiPolygon,
+    GeometryCollection,
+}